Add lists of N-containing intervals
These are just hard-coded into the source code. This could be done in a
nicer way, but then we need to change the package structure. It’s nice
that this is just a single file at the moment.

We already have other constants (chromosome lengths, the CGH template)
in there as well.